questions 15-20, finished implementation

+++ mia_capoli_code/Java/Assigment1/src/main/java/com/revature/questions/Question15.java
@@ -5,22 +5,22 @@ import com.revature.interfaces.MathMethods;
 public class Question15 implements MathMethods {
 
     @Override
-    public int additionM(int a, int b) {
+    public double additionM(double a, double b) {
         return a + b;
     }
 
     @Override
-    public int subtractionM(int a, int b) {
+    public double subtractionM(double a, double b) {
         return a - b;
     }
 
     @Override
-    public int multiplicationM(int a, int b) {
+    public double multiplicationM(double a, double b) {
         return a * b;
     }
 
     @Override
-    public int divisionM(int a, int b) {
+    public double divisionM(double a, double b) {
         return a / b;
     }
 }
